Z-Index refactor
This commit aims to standardize/fix most of the z-index issues present
in the app. By and large, cards should stay on top when moving around
(dragging or programatically), and all of the modals follow the max value
of 200.

I was very surprised to find seemingly inexplicable missing logic in the
context and <Card /> component, hopefully this covers most of the edge
cases.