Major Mobile Refactor
This commit brings full-fledged support for ui/ux on mobile devices. There
will certainly be more edge cases to rough out, however with the addition
of the <Drawer /> component and a general migration to mobile friendly
components, this commit is a major step forward in the right direction,
potentially towards a mobile app?