Another combined refactor
Fixed logic for rotating decks after a successful vote,
previously it didn't merge the current cards in the player's
hand back into the deck before rotating the deck, effectively
removing the player's cards from the game.

Scoreboard no longer automatically proceeds after ~15s, now waits for
the room's host to click the "Start next Round"/"Return to room" button.
Along with this, we removed all of the logic surrounding the (now unnecessary)
"playerIDToStartNextRound" state.

Also fixed incorrect logic for determining whether a game was currently
in progress or not in leaveRoomHandler

Small semanic changes to the tutorial modal.